Move base classes
