+++ kornia/augmentation/_2d/geometric/__init__.py
@@ -6,7 +6,6 @@ from kornia.augmentation._2d.geometric.fisheye import RandomFisheye
 from kornia.augmentation._2d.geometric.horizontal_flip import RandomHorizontalFlip
 from kornia.augmentation._2d.geometric.pad import PadTo
 from kornia.augmentation._2d.geometric.perspective import RandomPerspective
-from kornia.augmentation._2d.geometric.posterize import RandomPosterize
 from kornia.augmentation._2d.geometric.resized_crop import RandomResizedCrop
 from kornia.augmentation._2d.geometric.rotation import RandomRotation
 from kornia.augmentation._2d.geometric.thin_plate_spline import RandomThinPlateSpline
