+++ kornia/augmentation/_2d/geometric/crop.py
@@ -4,8 +4,8 @@ import torch
 from torch.nn.functional import pad
 
 from kornia.augmentation import random_generator as rg
+from kornia.augmentation._2d.geometric.base import GeometricAugmentationBase2D
 from kornia.augmentation.base import TensorWithTransformMat
-from kornia.augmentation.base_2d import GeometricAugmentationBase2D
 from kornia.augmentation.utils import _transform_input
 from kornia.constants import Resample
 from kornia.geometry.transform import crop_by_transform_mat, get_perspective_transform
