+++ kornia/augmentation/_2d/intensity/color_jitter.py
@@ -3,7 +3,7 @@ from typing import Dict, List, Optional, Tuple, Union, cast
 import torch
 
 from kornia.augmentation import random_generator as rg
-from kornia.augmentation.base_2d import IntensityAugmentationBase2D
+from kornia.augmentation._2d.intensity.base import IntensityAugmentationBase2D
 from kornia.constants import pi
 from kornia.enhance import adjust_brightness, adjust_contrast, adjust_hue, adjust_saturation
