+++ kornia/augmentation/_2d/intensity/sharpness.py
@@ -3,7 +3,7 @@ from typing import Dict, Optional, Tuple, Union, cast
 import torch
 
 from kornia.augmentation import random_generator as rg
-from kornia.augmentation.base_2d import IntensityAugmentationBase2D
+from kornia.augmentation._2d.intensity.base import IntensityAugmentationBase2D
 from kornia.enhance import sharpness
