+++ kornia/augmentation/_2d/intensity/solarize.py
@@ -3,7 +3,7 @@ from typing import Dict, List, Optional, Tuple, Union, cast
 import torch
 
 from kornia.augmentation import random_generator as rg
-from kornia.augmentation.base_2d import IntensityAugmentationBase2D
+from kornia.augmentation._2d.intensity.base import IntensityAugmentationBase2D
 from kornia.enhance import solarize
