+++ kornia/augmentation/_2d/mix/cutmix.py
@@ -4,7 +4,7 @@ from typing import Dict, Optional, Tuple, Union, cast
 import torch
 
 from kornia.augmentation import random_generator as rg
-from kornia.augmentation.base_2d import MixAugmentationBase
+from kornia.augmentation._2d.mix.base import MixAugmentationBase
 from kornia.augmentation.utils import _shape_validation
 from kornia.geometry.bbox import bbox_to_mask, infer_bbox_shape
