+++ kornia/augmentation/_3d/geometric/perspective.py
@@ -3,7 +3,7 @@ from typing import Dict, Optional, Union, cast
 import torch
 
 from kornia.augmentation import random_generator as rg
-from kornia.augmentation.base_3d import AugmentationBase3D
+from kornia.augmentation._3d.base import AugmentationBase3D
 from kornia.constants import Resample
 from kornia.geometry import get_perspective_transform3d, warp_perspective3d
