+++ kornia/augmentation/_3d/geometric/rotation.py
@@ -3,7 +3,7 @@ from typing import Dict, Optional, Tuple, Union, cast
 import torch
 
 from kornia.augmentation import random_generator as rg
-from kornia.augmentation.base_3d import AugmentationBase3D
+from kornia.augmentation._3d.base import AugmentationBase3D
 from kornia.constants import Resample
 from kornia.geometry import affine3d
 from kornia.geometry.transform.affwarp import _compute_rotation_matrix3d, _compute_tensor_center3d
