+++ kornia/augmentation/__init__.py
@@ -29,6 +29,10 @@ from kornia.augmentation._2d import (
     RandomThinPlateSpline,
     RandomVerticalFlip,
 )
+from kornia.augmentation._2d.base import AugmentationBase2D
+from kornia.augmentation._2d.geometric.base import GeometricAugmentationBase2D
+from kornia.augmentation._2d.intensity.base import IntensityAugmentationBase2D
+from kornia.augmentation._2d.mix.base import MixAugmentationBase
 from kornia.augmentation._3d import (
     CenterCrop3D,
     RandomAffine3D,
@@ -41,12 +45,14 @@ from kornia.augmentation._3d import (
     RandomRotation3D,
     RandomVerticalFlip3D,
 )
-from kornia.augmentation.base_2d import AugmentationBase2D
-from kornia.augmentation.base_3d import AugmentationBase3D
+from kornia.augmentation._3d.base import AugmentationBase3D
 from kornia.augmentation.container import AugmentationSequential, ImageSequential, PatchSequential, VideoSequential
 
 __all__ = [
     "AugmentationBase2D",
+    "GeometricAugmentationBase2D",
+    "IntensityAugmentationBase2D",
+    "MixAugmentationBase",
     "CenterCrop",
     "ColorJitter",
     "Normalize",
