--- kornia/augmentation/base_2d.py
@@ -1,262 +1,0 @@
-import warnings
-from typing import Dict, Optional, Tuple, cast
-
-import torch
-
-import kornia
-from kornia.augmentation.base import TensorWithTransformMat, _AugmentationBase, _BasicAugmentationBase
-from kornia.augmentation.utils import _transform_input, _transform_output_shape, _validate_input_dtype
-from kornia.utils.helpers import _torch_inverse_cast
-
-
-class AugmentationBase2D(_AugmentationBase):
-    r"""AugmentationBase2D base class for customized augmentation implementations.
-
-    For any augmentation, the implementation of "generate_parameters" and "apply_transform" are required while the
-    "compute_transformation" is only required when passing "return_transform" as True.
-
-    Args:
-        p: probability for applying an augmentation. This param controls the augmentation probabilities
-          element-wise for a batch.
-        p_batch: probability for applying an augmentation to a batch. This param controls the augmentation
-          probabilities batch-wise.
-        return_transform: if ``True`` return the matrix describing the geometric transformation applied to each
-          input tensor. If ``False`` and the input is a tuple the applied transformation won't be concatenated.
-        same_on_batch: apply the same transformation across the batch.
-        keepdim: whether to keep the output shape the same as input ``True`` or broadcast it to the batch
-          form ``False``.
-    """
-
-    def __check_batching__(self, input: TensorWithTransformMat):
-        if isinstance(input, tuple):
-            inp, mat = input
-            if len(inp.shape) == 4:
-                if len(mat.shape) != 3:
-                    raise AssertionError('Input tensor is in batch mode ' 'but transformation matrix is not')
-                if mat.shape[0] != inp.shape[0]:
-                    raise AssertionError(
-                        f"In batch dimension, input has {inp.shape[0]} but transformation matrix has {mat.shape[0]}"
-                    )
-            elif len(inp.shape) in (2, 3):
-                if len(mat.shape) != 2:
-                    raise AssertionError("Input tensor is in non-batch mode but transformation matrix is not")
-            else:
-                raise ValueError(f'Unrecognized output shape. Expected 2, 3, or 4, got {len(inp.shape)}')
-
-    def transform_tensor(self, input: torch.Tensor) -> torch.Tensor:
-        """Convert any incoming (H, W), (C, H, W) and (B, C, H, W) into (B, C, H, W)."""
-        _validate_input_dtype(input, accepted_dtypes=[torch.float16, torch.float32, torch.float64])
-        return _transform_input(input)
-
-    def identity_matrix(self, input) -> torch.Tensor:
-        """Return 3x3 identity matrix."""
-        return kornia.eye_like(3, input)
-
-
-class GeometricAugmentationBase2D(AugmentationBase2D):
-    r"""GeometricAugmentationBase2D base class for customized geometric augmentation implementations.
-
-    For any augmentation, the implementation of "generate_parameters" and "apply_transform" are required while the
-    "compute_transformation" is only required when passing "return_transform" as True.
-
-    Args:
-        p: probability for applying an augmentation. This param controls the augmentation probabilities
-          element-wise for a batch.
-        p_batch: probability for applying an augmentation to a batch. This param controls the augmentation
-          probabilities batch-wise.
-        return_transform: if ``True`` return the matrix describing the geometric transformation applied to each
-          input tensor. If ``False`` and the input is a tuple the applied transformation won't be concatenated.
-        same_on_batch: apply the same transformation across the batch.
-        keepdim: whether to keep the output shape the same as input ``True`` or broadcast it
-          to the batch form ``False``.
-    """
-
-    def inverse_transform(
-        self,
-        input: torch.Tensor,
-        transform: Optional[torch.Tensor] = None,
-        size: Optional[Tuple[int, int]] = None,
-        **kwargs,
-    ) -> torch.Tensor:
-        """By default, the exact transformation as ``apply_transform`` will be used."""
-        raise NotImplementedError
-
-    def compute_inverse_transformation(self, transform: torch.Tensor):
-        """Compute the inverse transform of given transformation matrices."""
-        return _torch_inverse_cast(transform)
-
-    def get_transformation_matrix(
-        self, input: torch.Tensor, params: Optional[Dict[str, torch.Tensor]] = None
-    ) -> torch.Tensor:
-        if params is not None:
-            transform = self.compute_transformation(input[params['batch_prob']], params)
-        elif not hasattr(self, "_transform_matrix"):
-            params = self.forward_parameters(input.shape)
-            transform = self.identity_matrix(input)
-            transform[params['batch_prob']] = self.compute_transformation(input[params['batch_prob']], params)
-        else:
-            transform = self._transform_matrix
-        return torch.as_tensor(transform, device=input.device, dtype=input.dtype)
-
-    def inverse(
-        self,
-        input: TensorWithTransformMat,
-        params: Optional[Dict[str, torch.Tensor]] = None,
-        size: Optional[Tuple[int, int]] = None,
-        **kwargs,
-    ) -> torch.Tensor:
-        if isinstance(input, (list, tuple)):
-            input, transform = input
-        else:
-            transform = self.get_transformation_matrix(input, params)
-        if params is not None:
-            transform = self.identity_matrix(input)
-            transform[params['batch_prob']] = self.compute_transformation(input[params['batch_prob']], params)
-
-        ori_shape = input.shape
-        in_tensor = self.transform_tensor(input)
-        batch_shape = input.shape
-        if params is None:
-            params = self._params
-        if size is None and "forward_input_shape" in params:
-            # Majorly for cropping functions
-            size = params['forward_input_shape'].numpy().tolist()
-            size = (size[-2], size[-1])
-        if 'batch_prob' not in params:
-            params['batch_prob'] = torch.tensor([True] * batch_shape[0])
-            warnings.warn("`batch_prob` is not found in params. Will assume applying on all data.")
-        output = input.clone()
-        to_apply = params['batch_prob']
-        # if no augmentation needed
-        if torch.sum(to_apply) == 0:
-            output = in_tensor
-        # if all data needs to be augmented
-        elif torch.sum(to_apply) == len(to_apply):
-            transform = self.compute_inverse_transformation(transform)
-            output = self.inverse_transform(in_tensor, transform, size, **kwargs)
-        else:
-            transform[to_apply] = self.compute_inverse_transformation(transform[to_apply])
-            output[to_apply] = self.inverse_transform(in_tensor[to_apply], transform[to_apply], size, **kwargs)
-        return cast(torch.Tensor, _transform_output_shape(output, ori_shape)) if self.keepdim else output
-
-
-class IntensityAugmentationBase2D(AugmentationBase2D):
-    r"""IntensityAugmentationBase2D base class for customized intensity augmentation implementations.
-
-    For any augmentation, the implementation of "generate_parameters" and "apply_transform" are required while the
-    "compute_transformation" is only required when passing "return_transform" as True.
-
-    Args:
-        p: probability for applying an augmentation. This param controls the augmentation probabilities
-          element-wise for a batch.
-        p_batch: probability for applying an augmentation to a batch. This param controls the augmentation
-          probabilities batch-wise.
-        return_transform: if ``True`` return the matrix describing the geometric transformation applied to each
-          input tensor. If ``False`` and the input is a tuple the applied transformation  won't be concatenated.
-        same_on_batch: apply the same transformation across the batch.
-        keepdim: whether to keep the output shape the same as input ``True`` or broadcast it
-          to the batch form ``False``.
-    """
-
-    def compute_transformation(self, input: torch.Tensor, params: Dict[str, torch.Tensor]) -> torch.Tensor:
-        return self.identity_matrix(input)
-
-
-class MixAugmentationBase(_BasicAugmentationBase):
-    r"""MixAugmentationBase base class for customized mix augmentation implementations.
-
-    For any augmentation, the implementation of "generate_parameters" and "apply_transform" are required.
-    "apply_transform" will need to handle the probabilities internally.
-
-    Args:
-        p: probability for applying an augmentation. This param controls if to apply the augmentation for the batch.
-        p_batch: probability for applying an augmentation to a batch. This param controls the augmentation
-          probabilities batch-wise.
-        same_on_batch: apply the same transformation across the batch.
-        keepdim: whether to keep the output shape the same as input ``True`` or broadcast it
-          to the batch form ``False``.
-    """
-
-    def __init__(self, p: float, p_batch: float, same_on_batch: bool = False, keepdim: bool = False) -> None:
-        super().__init__(p, p_batch=p_batch, same_on_batch=same_on_batch, keepdim=keepdim)
-
-    def __check_batching__(self, input: TensorWithTransformMat):
-        if isinstance(input, tuple):
-            inp, mat = input
-            if len(inp.shape) == 4:
-                if len(mat.shape) != 3:
-                    raise AssertionError('Input tensor is in batch mode ' 'but transformation matrix is not')
-                if mat.shape[0] != inp.shape[0]:
-                    raise AssertionError(
-                        f"In batch dimension, input has {inp.shape[0]} but transformation matrix has {mat.shape[0]}"
-                    )
-            elif len(inp.shape) in (2, 3):
-                if len(mat.shape) != 2:
-                    raise AssertionError("Input tensor is in non-batch mode but transformation matrix is not")
-            else:
-                raise ValueError(f'Unrecognized output shape. Expected 2, 3, or 4, got {len(inp.shape)}')
-
-    def __unpack_input__(  # type: ignore
-        self, input: TensorWithTransformMat
-    ) -> Tuple[torch.Tensor, Optional[torch.Tensor]]:
-        if isinstance(input, tuple):
-            in_tensor = input[0]
-            in_transformation = input[1]
-            return in_tensor, in_transformation
-        in_tensor = input
-        return in_tensor, None
-
-    def transform_tensor(self, input: torch.Tensor) -> torch.Tensor:
-        """Convert any incoming (H, W), (C, H, W) and (B, C, H, W) into (B, C, H, W)."""
-        _validate_input_dtype(input, accepted_dtypes=[torch.float16, torch.float32, torch.float64])
-        return _transform_input(input)
-
-    def apply_transform(  # type: ignore
-        self, input: torch.Tensor, label: torch.Tensor, params: Dict[str, torch.Tensor]
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
-        raise NotImplementedError
-
-    def apply_func(  # type: ignore
-        self, in_tensor: torch.Tensor, label: torch.Tensor, params: Dict[str, torch.Tensor]
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
-        to_apply = params['batch_prob']
-
-        # if no augmentation needed
-        if torch.sum(to_apply) == 0:
-            output = in_tensor
-        # if all data needs to be augmented
-        elif torch.sum(to_apply) == len(to_apply):
-            output, label = self.apply_transform(in_tensor, label, params)
-        else:
-            raise ValueError(
-                "Mix augmentations must be performed batch-wisely. Element-wise augmentation is not supported."
-            )
-
-        return output, label
-
-    def forward(  # type: ignore
-        self,
-        input: TensorWithTransformMat,
-        label: Optional[torch.Tensor] = None,
-        params: Optional[Dict[str, torch.Tensor]] = None,
-    ) -> Tuple[TensorWithTransformMat, torch.Tensor]:
-        in_tensor, in_trans = self.__unpack_input__(input)
-        ori_shape = in_tensor.shape
-        in_tensor = self.transform_tensor(in_tensor)
-        # If label is not provided, it would output the indices instead.
-        if label is None:
-            if isinstance(input, (tuple, list)):
-                device = input[0].device
-            else:
-                device = input.device
-            label = torch.arange(0, in_tensor.size(0), device=device, dtype=torch.long)
-        if params is None:
-            batch_shape = in_tensor.shape
-            params = self.forward_parameters(batch_shape)
-        self._params = params
-
-        output, lab = self.apply_func(in_tensor, label, self._params)
-        output = _transform_output_shape(output, ori_shape) if self.keepdim else output  # type: ignore
-        if in_trans is not None:
-            return (output, in_trans), lab
-        return output, lab
