+++ kornia/augmentation/container/video.py
@@ -5,8 +5,8 @@ import torch.nn as nn
 
 import kornia
 from kornia.augmentation import RandomCrop
+from kornia.augmentation._2d.base_2d import MixAugmentationBase
 from kornia.augmentation.base import TensorWithTransformMat, _AugmentationBase
-from kornia.augmentation.base_2d import MixAugmentationBase
 from kornia.augmentation.container.base import SequentialBase
 from kornia.augmentation.container.image import ImageSequential, ParamItem
 from kornia.augmentation.container.utils import InputApplyInverse, MaskApplyInverse
