+++ test/augmentation/test_augmentation.py
@@ -30,7 +30,7 @@ from kornia.augmentation import (
     RandomThinPlateSpline,
     RandomVerticalFlip,
 )
-from kornia.augmentation.base_2d import AugmentationBase2D
+from kornia.augmentation._2d.base import AugmentationBase2D
 from kornia.constants import Resample, pi
 from kornia.geometry import transform_points
 from kornia.testing import BaseTester, assert_close, default_with_one_parameter_changed
@@ -144,13 +144,15 @@ class CommonTests(BaseTester):
         expected_transformation_shape = torch.Size((generated_params['batch_prob'].sum(), 3, 3))
         test_input = torch.ones(batch_shape, device=self.device, dtype=self.dtype)
         transformation = augmentation.compute_transformation(
-            test_input[generated_params['batch_prob']], generated_params)
+            test_input[generated_params['batch_prob']], generated_params
+        )
         assert transformation.shape == expected_transformation_shape
 
         # apply_transform can be called and returns the correct batch sized output
         if generated_params['batch_prob'].sum() != 0:
             output = augmentation.apply_transform(
-                test_input[generated_params['batch_prob']], generated_params, transformation)
+                test_input[generated_params['batch_prob']], generated_params, transformation
+            )
             assert output.shape[0] == generated_params['batch_prob'].sum()
         else:
             # Re-generate parameters if 0 batch size
